+++ measure/measure.h
@@ -2,7 +2,7 @@
 #define MEASURE_H
 
 #include <ostream>
-#include "tbb/tick_count.h"
+#include <tbb/tick_count.h>
 
 typedef tbb::tick_count tbb_time_t;
 
@@ -10,6 +10,6 @@ tbb_time_t gettime ();
 
 double time_diff (tbb_time_t start, tbb_time_t end);
 
-void measure (std::ostream& result, int iterations, void (*F)());
+void measure (std::ostream& result, size_t iterations, void (*F)());
 
 #endif
